Computes the crossing survival times

Computes the crossing survival times

Usage

# S3 method for ypreg
cross_time(object, newdata1, newdata2, conf.level = 0.95, nboot = 1000, ...)

Arguments

object

an object of class ypreg

newdata1

a data frame containing the first set of explanatory variables

newdata2

a data frame containing the second set of explanatory variables

conf.level

level of the confidence/credible intervals

nboot

number of bootstrap samples (default nboot=1000).

...

further arguments passed to or from other methods.

Value

the crossing survival time

Examples

# \donttest{
library(survstan)
data(ipass)
fit <- ypreg(Surv(time, status)~arm, data=ipass, baseline = "weibull")
summary(fit)
#> Call:
#> ypreg(formula = Surv(time, status) ~ arm, data = ipass, baseline = "weibull")
#> 
#> Yang & Prentice model fit with weibull baseline distribution 
#> 
#> Regression coefficients:
#>            Estimate    StdErr  z.value   p.value    
#> short-arm  1.361063  0.182368   7.4633  8.44e-14 ***
#> long-arm  -1.365391  0.082943 -16.4619 < 2.2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Baseline parameters:
#>       estimate       se     2.5%  97.5%
#> alpha 1.846062 0.032814 1.781747 1.9104
#> gamma 6.961753 0.501101 5.979614 7.9439
#> --- 
#> loglik = -2772.375   AIC = 5552.751 
newdata1 <- data.frame(arm=0)
newdata2 <- data.frame(arm=1)
tcross <- cross_time(fit, newdata1, newdata2, nboot = 10)
tcross
#>       Est.     2.5%    97.5%
#> 1 5.867165 5.371126 6.514087
# }