The 'survstan' package.

The aim of the R package survstan is to provide a toolkit for fitting survival models using Stan. The R package survstan can be used to fit right-censored survival data under independent censoring. The implemented models allow the fitting of survival data in the presence/absence of covariates. All inferential procedures are currently based on the maximum likelihood (ML) approach.

_PACKAGE

References

Stan Development Team (2023). “RStan: the R interface to Stan.” R package version 2.21.8, https://mc-stan.org/.

Lawless JF (2002). Statistical Models and Methods for Lifetime Data, Wiley Series in Probability and Statistics, 2nd Edition edition. John Wiley and Sons. ISBN 9780471372158.

Bennett S (1983). “Analysis of survival data by the proportional odds model.” Statistics in Medicine, 2(2), 273-277. doi:10.1002/sim.4780020223 .

Chen YQ, Wang M (2000). “Analysis of Accelerated Hazards Models.” Journal of the American Statistical Association, 95(450), 608-618. doi:10.1080/01621459.2000.10474236 .

Demarqui FN, Mayrink VD (2021). “Yang and Prentice model with piecewise exponential baseline distribution for modeling lifetime data with crossing survival curves.” Brazilian Journal of Probability and Statistics, 35(1), 172 – 186. doi:10.1214/20-BJPS471 .